requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
langchain==1.0.0a15
langchain-core==1.0.0rc1
langgraph==1.0.0a4
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.7.0a2
langgraph-sdk==0.2.9
langsmith==0.4.37
